Attempting to fix the issue where the same collision can be triggered multiple times in the same frame
Former-commit-id: 604104ca9058f3f2c3396ba38eaddf406ab111a8 [formerly 979005fc8884063b34413e907ed9c339c0f197ab [formerly 98876237ef3d758c466d3530568c8dcc6026ac69]]
Former-commit-id: 58efadb4cd2ce2b058eaf7ee2c7328acdb234c19
Former-commit-id: 1f2e2d342639ec4ff29a0260a79e3509d6f049f0

+++ BrickwareMath/src/Matrix3.cpp
@@ -65,14 +65,14 @@ Matrix3 Matrix3::getInverse()
 {
 	Matrix3 inverse;
 
-	double determinant = (matrix[0][0] * ((matrix[1][1] * matrix[2][2]) - (matrix[2][1] * matrix[1][2])))
+	float determinant = (matrix[0][0] * ((matrix[1][1] * matrix[2][2]) - (matrix[2][1] * matrix[1][2])))
 					  - (matrix[0][1] * ((matrix[1][0] * matrix[2][2]) - (matrix[1][2] * matrix[2][0])))
 					  + (matrix[0][2] * ((matrix[1][0] * matrix[2][1]) - (matrix[1][1] * matrix[2][0])));
 
 	//Return an identity matrix if there is no inverse
 	if (determinant > 0)
 	{
-		double inverseDeterminant = 1 / determinant;
+		float inverseDeterminant = 1 / determinant;
 
 		inverse[0][0] =  ((matrix[1][1] * matrix[2][2]) - (matrix[2][1] * matrix[1][2])) * inverseDeterminant;
 		inverse[0][1] =  ((matrix[0][2] * matrix[2][1]) - (matrix[0][1] * matrix[2][2])) * inverseDeterminant;
